# ARROW-11042: [Rust][DataFusion] Increase default batch size